backend/requirements.txt
fastapi
uvicorn
pydantic>=2.0
pydantic-settings
google-cloud-discoveryengine
google-cloud-aiplatform
firebase-admin
sse-starlette
python-dotenv
genkit
genkit[google-genai]